+++ packages/query/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow restrictions on array based queries, with flag `--query-batch-limit` (#2172)
 
 ## [2.7.0] - 2023-11-15
 ### Added